[be] refactor: multiResponsedto 추가
