+++ server/src/main/java/com/CreatorConnect/server/response/PageInfo.java
@@ -1,0 +1,13 @@
+package com.CreatorConnect.server.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class PageInfo {
+    private int page;
+    private int size;
+    private long totalElements;
+    private int totalPages;
+}
