+++ server/src/main/java/com/CreatorConnect/server/response/SingleResponseDto.java
@@ -1,0 +1,10 @@
+package com.CreatorConnect.server.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class SingleResponseDto<T> {
+    private T data;
+}
